didnt attach everything last time.
