+++ Assets/Resources/UI/MenuLeft/MenuLeft.prefab.meta
