--- Assets/Scripts/Models/Character/Health.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a2ecfcb201ca2cf4f9f81a790ba1f6e4
-timeCreated: 1475704393
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
